3619: remove use of older hardcoded const, and use backend as src of truth. use new SHA for openapi. update naming and featureFlagging.
